feat: add Linux/Raspberry Pi platform port for E22-900M30S (SX1262)
Adds a complete build target for running RIVR on a Raspberry Pi with an
EBYTE E22-900M30S (SX1262) LoRa module, connected over SPI.

GPIO wiring (BCM numbers):
  NSS=21  BUSY=20  NRST=18  DIO1=16  RXEN=12  TXEN=13
  MISO=9  MOSI=10  SCK=11  → /dev/spidev0.0

New files:
- firmware_core/platform_linux.{h,c}       — SPI (spidev) + GPIO (libgpiod) HAL
- firmware_core/radio_sx1262_linux.c        — SX1262 driver (pthread mutex, atomic DIO1)
- firmware_core/timebase_linux.c            — 1 ms pthread monotonic timer
- firmware_core/rivr_panic_linux.c          — panic handler (fprintf + abort)
- firmware_core/main_linux.c                — main(), MAC-derived node ID, main loop
- firmware_core/linux/include/              — ESP-IDF compatibility stubs
- Makefile.linux                            — build system (gcc + cargo)

Build:
  sudo apt install libgpiod-dev
  # enable SPI: dtparam=spi=on in /boot/firmware/config.txt
  make -f Makefile.linux
  sudo ./rivr

No BLE, no display, no USB bridge. NVS stubbed → compiled-in default program.

+++ rivr_layer/rivr_sinks.c
@@ -76,7 +76,7 @@ void rf_tx_sink_cb(const rivr_value_t *v, void *user_ctx)
         req.data[1] = (uint8_t)(lmp & 0xFFu);
         req.data[2] = (uint8_t)(lmp >> 8u);
 
-        uint8_t plen = (slen >= payload_offset) ? (slen - payload_offset) : 0u;
+        uint8_t plen = (slen >= payload_offset) ? (uint8_t)(slen - payload_offset) : 0u;
         if (plen > RF_MAX_PAYLOAD_LEN - 3u) {
             plen = RF_MAX_PAYLOAD_LEN - 3u;
             ESP_LOGW(TAG, "rf_tx: payload truncated to %u bytes", plen);
